requirements.txt
Django==3.0.2
django-cors-headers==3.1.1
django-webpack-loader==0.6.0
djangorestframework==3.11.0
django-rest-auth==0.9.5
git+https://github.com/SergioVenicio21/eyeD3.git
mutagen==1.42.0
pathlib==1.0.1
Pillow==8.0.0
python-magic==0.4.15
pytz==2019.3
six==1.14.0
python-decouple==3.3
gunicorn==20.0.4
Collectfast==1.3.1
dj-database-url==0.5.0
psycopg2-binary==2.8.4
